+++ components/ordered-unordered-list/ordered-unordered-list.stories.tsx
@@ -1,4 +1,4 @@
-import type { Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from '@storybook/react-webpack5';
 import { defaultArgs } from './src/defaultArgs';
 import { List } from './src/template';