+++ documentation/demopages/experimenteel/deRegelmaat/Profiel.stories.tsx
@@ -1,4 +1,4 @@
-import type { Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from '@storybook/react-webpack5';
 import Profiel from './Profiel';
 
 const meta: Meta<typeof Profiel> = {