+++ documentation/demopages/experimenteel/jongeVissers/KostenOpvoeren.stories.tsx
@@ -1,4 +1,4 @@
-import type { Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from '@storybook/react-webpack5';
 import KostenOpvoeren from './KostenOpvoeren';
 
 const meta: Meta<typeof KostenOpvoeren> = {