+++ documentation/demopages/voorbeelden/algemeen/Report.stories.tsx
@@ -1,4 +1,4 @@
-import type { Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from '@storybook/react-webpack5';
 import Report from './Report';
 
 const meta: Meta<typeof Report> = {