+++ components/dialog/CHANGELOG.md
@@ -1,5 +1,11 @@
 # `@nl-rvo/css-dialog`
 
+## 1.4.4
+
+### Patch Changes
+
+- 11eb258: Update imports to support storybook v10
+
 ## 1.4.3
 
 ### Patch Changes