+++ components/heading/CHANGELOG.md
@@ -1,5 +1,11 @@
 # `@nl-rvo/css-heading`
 
+## 2.1.1
+
+### Patch Changes
+
+- 11eb258: Update imports to support storybook v10
+
 ## 2.1.0
 
 ### Minor Changes