+++ components/image/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nl-rvo/css-image
 
+## 1.1.1
+
+### Patch Changes
+
+- 11eb258: Update imports to support storybook v10
+
 ## 1.1.0
 
 ### Minor Changes