+++ components/skeleton/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Pre-releases
 
+## 1.0.2
+
+### Patch Changes
+
+- 11eb258: Update imports to support storybook v10
+
 ## 1.0.1
 
 ### Patch Changes